pyproject.toml
[tool.poetry]
name = "stock service"
version = "0.1.0"
description = "An api for the stock market"
authors = ["Mauro J. Lopez"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
psycopg2 = "^2.9.2"
djangorestframework = "^3.12.4"
djangorestframework-api-key = "2."

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
